requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
attrs==19.1.0             # via jsonschema
beautifulsoup4==4.8.0
boto3==1.9.232
botocore==1.12.233        # via boto3, s3transfer
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
dateparser==0.7.0
directory-ch-client==2.1.0
directory-client-core==6.1.0  # via directory-ch-client, directory-forms-api-client
directory-components==34.0.1
directory-constants==20.3.0  # via directory-components
directory-forms-api-client==5.0.0
directory-healthcheck==1.1.2
django-environ==0.4.5
django-formtools==2.1
django-health-check==3.8.0  # via directory-healthcheck
django-recaptcha==2.0.5
django-redis==4.10.0
django-storages==1.7.2
django==2.2.5
djangorestframework==3.10.3
docutils==0.15.2          # via botocore
gunicorn==19.9.0
idna==2.8                 # via requests
jmespath==0.9.4           # via boto3, botocore
jsonschema==3.0.2         # via directory-components
mohawk==0.3.4             # via sigauth
monotonic==1.5            # via directory-ch-client, directory-client-core
pyrsistent==0.15.4        # via jsonschema
python-dateutil==2.8.0    # via botocore, dateparser
pytz==2019.2              # via dateparser, django, tzlocal
raven==6.9.0
redis==3.3.8              # via django-redis
regex==2019.8.19          # via dateparser
requests==2.21.0
s3transfer==0.2.1         # via boto3
sigauth==4.1.1            # via directory-client-core
six==1.12.0               # via jsonschema, mohawk, pyrsistent, python-dateutil, w3lib
soupsieve==1.9.3          # via beautifulsoup4
sqlparse==0.3.0           # via django
tzlocal==2.0.0            # via dateparser
urllib3==1.24.3
w3lib==1.21.0             # via directory-client-core
whitenoise==4.1.3

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via jsonschema
